#16674 Change ResultSetDataContainerOptions.selectedRows type to Integer

+++ plugins/org.jkiss.dbeaver.ui.editors.data/src/org/jkiss/dbeaver/ui/controls/resultset/ResultSetDataContainer.java
@@ -104,7 +104,7 @@ public class ResultSetDataContainer implements DBSDataContainer, DBPContextProvi
             try {
                 dataReceiver.fetchStart(session, resultSet, firstRow, maxRows);
                 while (!session.getProgressMonitor().isCanceled() && resultSet.nextRow()) {
-                    if (!proceedSelectedRowsOnly(flags) || options.getSelectedRows().contains((long) resultSet.curRow.getRowNumber())) {
+                    if (!proceedSelectedRowsOnly(flags) || options.getSelectedRows().contains(resultSet.curRow.getRowNumber())) {
                         dataReceiver.fetchRow(session, resultSet);
                     }
                     resultCount++;
